\name{print.PP}

\alias{print.PP}
\alias{print.PPA}

\title{
"PP" and "PPA" methods for generic function print()
}

\description{
Nicely print or display a "PP" or "PPA" object.
}

\usage{
\method{print}{PP}(x, \dots)

\method{print}{PPA}(x, \dots)
}

\arguments{
  \item{x}{An object that inherits "PP" class, typically a "PP" or "PPA" object.}
  \item{\dots}{Not used by the method.}
}

\value{
These methods have no returned values.
}

\author{
Zheyuan Li \email{zheyuan.li@bath.edu}
}
